# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bứt Phá Marketing ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: loginAuth
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    type: object

  # ── DEBTS ────────────────────────────────────────────────────────────
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      parameters:
        - in: query
          name: customerId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Debt" }
                  pagination: { $ref: "#/components/schemas/Pagination" }
    post:
      operationId: createDebt
      tags: [debts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DebtInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Debt" }
  /debts/{id}:
    get:
      operationId: getDebt
      tags: [debts]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Debt" }
    patch:
      operationId: updateDebt
      tags: [debts]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DebtInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Debt" }
    delete:
      operationId: deleteDebt
      tags: [debts]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No content

  # ── DOMAINS ─────────────────────────────────────────────────────────
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: customerId
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Domain" }
                  pagination: { $ref: "#/components/schemas/Pagination" }
    post:
      operationId: createDomain
      tags: [domains]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DomainInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
  /domains/{id}:
    get:
      operationId: getDomain
      tags: [domains]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
    patch:
      operationId: updateDomain
      tags: [domains]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DomainInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
    delete:
      operationId: deleteDomain
      tags: [domains]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No content

  # ── HOSTINGS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    # ── DEBT ─────────────────────────────────────────────────────────────
    Debt:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        contractId: { type: ["integer", "null"] }
        totalAmount: { type: number }
        paidAmount: { type: number }
        dueDate: { type: string }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    DebtInput:
      type: object
      properties:
        customerId: { type: integer }
        contractId: { type: integer }
        totalAmount: { type: number }
        paidAmount: { type: number }
        dueDate: { type: string }
        status: { type: string }

    # ── DOMAIN ───────────────────────────────────────────────────────────
    Domain:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        domainName: { type: string }
        provider: { type: string }
        registerDate: { type: string }
        expireDate: { type: string }
        buyPrice: { type: ["number", "null"] }
        sellPrice: { type: ["number", "null"] }
        daysUntilExpiry: { type: integer }
        status: { type: string }
        note: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    DomainInput:
      type: object
      properties:
        customerId: { type: integer }
        domainName: { type: string }
        provider: { type: string }
        registerDate: { type: string }
        expireDate: { type: string }
        buyPrice: { type: number }
        sellPrice: { type: number }
        status: { type: string }
        note: { type: string }

    # ── HOSTING ──────────────────────────────────────────────────────────
...