# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bứt Phá Marketing ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface Domain {
  id?: number;
  customerId?: number;
  /** @nullable */
  customerName?: string | null;
  domainName?: string;
  provider?: string;
  registerDate?: string;
  expireDate?: string;
  /** @nullable */
  buyPrice?: number | null;
  /** @nullable */
  sellPrice?: number | null;
  daysUntilExpiry?: number;
  status?: string;
  /** @nullable */
  note?: string | null;
  createdAt?: string;
  updatedAt?: string;
}